+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Tree Stump feature in Forest Lands
+
+### Contributors for this release
+
+- ScarabOasis, Dweblenod
+
 ## [1.20.1-1.11.2.0] - 2024-01-13
 
 ### Added